backend/requirements.txt
flask
flask-cors
redis
kafka-python
mysql-connector-python 